+++ vortex-vector/src/bool/mod.rs
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
+//! Definition and implementation of [`BoolVector`] and [`BoolVectorMut`].
+
 mod vector;
 pub use vector::BoolVector;
