+++ vortex-vector/src/vector/mod.rs
@@ -1,6 +1,9 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
+//! Definition of the [`Vector`] and [`VectorMut`] types, which represent fully decompressed
+//! (canonical) array data.
+
 use vortex_dtype::DType;
 use vortex_error::vortex_panic;
 
@@ -8,10 +11,24 @@ use crate::{
     BoolVector, BoolVectorMut, NullVector, NullVectorMut, PrimitiveVector, PrimitiveVectorMut,
 };
 
+/// Helper macros for working with the different variants of [`Vector`] and [`VectorMut`].
+///
+/// All macros are exported at the crate level with `#[macro_use]`.
 mod macros;
+
+/// Definition and implementation of [`VectorOps`](ops::VectorOps) and
+/// [`VectorMutOps`](ops::VectorMutOps) for [`Vector`] and [`VectorMut`], respecitively.
 pub(super) mod ops;
 
-/// An enum over all vector types.
+/// An enum over all kinds of immutable vectors, which represent fully decompressed (canonical)
+/// array data.
+///
+/// Most of the behavior of `Vector` is described by the [`VectorOps`] trait.
+///
+/// The mutable equivalent of this type is [`VectorMut`], which implements.
+///
+/// [`VectorOps`]: crate::VectorOps
+#[derive(Debug, Clone)]
 pub enum Vector {
     /// Null
     Null(NullVector),
@@ -19,7 +36,8 @@ pub enum Vector {
     Bool(BoolVector),
     /// Primitive
     ///
-    /// TODO(connor): Document that this is an enum, not a struct.
+    /// TODO(connor): Document that this is an enum, not a struct (to represent all possible
+    /// primitive native generics).
     Primitive(PrimitiveVector),
     // Decimal
     // Decimal(DecimalVector),
@@ -37,7 +55,15 @@ pub enum Vector {
     // Extension(ExtensionVector),
 }
 
-/// An enum over all mutable vector types.
+/// An enum over all kinds of mutable vectors, which represent fully decompressed (canonical) array
+/// data.
+///
+/// Most of the behavior of `VectorMut` is described by the [`VectorMutOps`] trait.
+///
+/// The immutable equivalent of this type is [`Vector`].
+///
+/// [`VectorMutOps`]: crate::VectorMutOps
+#[derive(Debug, Clone)]
 pub enum VectorMut {
     /// Null
     Null(NullVectorMut),
@@ -51,7 +77,7 @@ impl VectorMut {
     /// Create a new mutable vector with the given capacity and dtype.
     pub fn with_capacity(capacity: usize, dtype: &DType) -> Self {
         match dtype {
-            DType::Null => NullVectorMut::new(0).into(),
+            DType::Null => NullVectorMut::new(0).into(), // `NullVector` has `usize::MAX` capacity.
             DType::Bool(n) => BoolVectorMut::with_capacity(capacity, *n).into(),
             DType::Primitive(ptype, nullability) => {
                 PrimitiveVectorMut::with_capacity(capacity, *ptype, *nullability).into()
